requirements.txt
streamlit==1.38.0
supabase==2.7.1
python-dotenv==1.0.1
schedule==1.2.2